rename `doc` directory to `docs`
There is a basic convention, established by the Developer Center
orchestration, that assumes that any component, so any repository,
will manage the contents to be fetched by the Developer
Center under the `docs` repository.

We have then to change the initial convention we followed here,
according to npm rules, in order to integrate this component
into the Developer Center.