fix(styles): Separate body and app classes to prevent breaking the `flex` chain. #409
BREAKING CHANGE: This PR splits the `as-app` class into `as-app-body` and `as-app`classes.
`as-app-body` is meant to be applied to the `body` document tag.
`as-app` is meant to be applied to the node that contains the whole application made with Airship.

Previous to this, since the app needs typically more than one node (one for the toolbar, one for the content), when used in a framework like Vue those two nodes needed to be wrapped in one div. Since components don't create the body tag, it caused an extra `div` that broke the display setting for the app and the layout broke.