fix(styles): refactor footer containers #502
BREAKING CHANGE: footer containers were introducing a lot of problems when mixing fixed and scrollable containers.
We've dropped `as-container` from the footer styles. Now, every content added to a `as-footer` is recommended to be wrapped by an element with class `as-box`.

Please refer to the layout documentation to know more about how to add content to both sidebars and footer.